% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_deepdives_multiple_methods.R
\name{country_dd_plot}
\alias{country_dd_plot}
\title{Wrapper for scrolly plots that do a deepdive on all methods}
\usage{
country_dd_plot(step, plots)
}
\arguments{
\item{step}{one of "s1" … "s5"}

\item{plots}{named list created by `plot_country_deepdives_multiple_methods()`}
}
\value{
a ggplot object
}
\description{
Wrapper for scrolly plots that do a deepdive on all methods
}
